Use json-to-dhall to implement yaml-to-dhall
