+++ dhall-json/dhall-json.cabal
@@ -9,10 +9,10 @@ Copyright: 2017 Gabriel Gonzalez
 Author: Gabriel Gonzalez
 Maintainer: Gabriel439@gmail.com
 Bug-Reports: https://github.com/dhall-lang/dhall-haskell/issues
-Synopsis: Compile Dhall to JSON or YAML
+Synopsis: Convert between Dhall and JSON or YAML
 Description:
-    Use this package if you want to compile Dhall expressions to JSON or YAML.
-    You can use this package as a library or an executable:
+    Use this package if you want to convert between Dhall expressions and JSON
+    or YAML. You can use this package as a library or an executable:
     .
     * See the "Dhall.JSON" module if you want to use this package as a library
     .
@@ -40,13 +40,18 @@ Library
         base                      >= 4.8.0.0  && < 5   ,
         aeson                     >= 1.0.0.0  && < 1.5 ,
         bytestring                               < 0.11,
+        containers                                     ,
         dhall                     >= 1.22.0   && < 1.24,
+        exceptions                >= 0.8.3    && < 0.11,
         lens                      >= 2.5      && < 4.18,
         optparse-applicative      >= 0.14.0.0 && < 0.15,
+        scientific                >= 0.3.0.0  && < 0.4 ,
         text                      >= 0.11.1.0 && < 1.3 ,
         unordered-containers                     < 0.3 ,
         vector
-    Exposed-Modules: Dhall.JSON
+    Exposed-Modules:
+        Dhall.JSON
+        Dhall.JSONToDhall
     GHC-Options: -Wall
 
     if flag(yaml-pre-0_11)
@@ -95,17 +100,41 @@ Executable json-to-dhall
         aeson-pretty                        < 0.9  ,
         bytestring                          < 0.11 ,
         dhall                                      ,
-        optparse-applicative                       ,
-        text                                < 1.3  ,
-        scientific           >= 0.3.0.0  && < 0.4  ,
+        dhall-json                                 ,
         exceptions           >= 0.8.3    && < 0.11 ,
-        containers                                 ,
-        unordered-containers >= 0.1.3.0  && < 0.3
+        optparse-applicative                       ,
+        text                                < 1.3
     if !impl(ghc >= 8.0) && !impl(eta >= 0.8.4)
       Build-Depends: semigroups == 0.18.*
     Other-Modules:
         Paths_dhall_json
     GHC-Options: -Wall
+
+Executable yaml-to-dhall
+    Hs-Source-Dirs: yaml-to-dhall
+    Main-Is: Main.hs
+    Build-Depends:
+        base                                       ,
+        aeson                                      ,
+        bytestring                          < 0.11 ,
+        dhall                                      ,
+        dhall-json                                 ,
+        exceptions           >= 0.8.3    && < 0.11 ,
+        optparse-applicative                       ,
+        text                                < 1.3
+    if !impl(ghc >= 8.0) && !impl(eta >= 0.8.4)
+      Build-Depends: semigroups == 0.18.*
+    Other-Modules:
+        Paths_dhall_json
+    GHC-Options: -Wall
+
+    if flag(yaml-pre-0_11)
+        Build-Depends:
+            yaml >= 0.5.0 && < 0.11
+    else
+        Build-Depends:
+            libyaml >= 0.1.1.0 && < 0.2 ,
+            yaml    >= 0.11.0  && < 0.12
 
 Test-Suite tasty
     Type: exitcode-stdio-1.0
