+++ dhall-json/examples/bower.dhall
@@ -1,7 +1,10 @@
 -- Using this file:
 -- json-to-dhall ./bower.dhall < ./bower.json | dhall
+-- yaml-to-dhall ./bower.dhall < ./bower.yaml | dhall
+--
 -- Round-trip test:
 -- json-to-dhall ./bower.dhall < ./bower.json | dhall | dhall-to-json
+-- yaml-to-dhall ./bower.dhall < ./bower.yaml | dhall | dhall-to-yaml
 
 { name            : Text
 , description     : Text
